requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
pydantic-settings==2.11.0
openai==1.108.1
httpx==0.28.1
python-multipart==0.0.20
